Refactor home package to view
