+++ TwoPage/src/main/java/com/microsoft/device/display/samples/twopage/MainActivity.kt
@@ -8,7 +8,7 @@ package com.microsoft.device.display.samples.twopage
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
-import com.microsoft.device.display.samples.twopage.ui.home.TwoPageApp
+import com.microsoft.device.display.samples.twopage.ui.view.TwoPageApp
 import com.microsoft.device.display.samples.twopage.ui.theme.TwoPageComposeSamplesTheme
 import com.microsoft.device.dualscreen.windowstate.WindowState
 import com.microsoft.device.dualscreen.windowstate.rememberWindowState
